refactor: `Client` internal module
