+++ src/Client.ts
@@ -5,18 +5,15 @@
  *
  * @since 1.0.0
  */
-import * as PlatformClient from "@effect/platform/Http/Client";
-import * as ClientRequest from "@effect/platform/Http/ClientRequest";
-import * as Schema from "@effect/schema/Schema";
-import * as Api from "effect-http/Api";
+import type * as ClientRequest from "@effect/platform/Http/ClientRequest";
+import type * as Schema from "@effect/schema/Schema";
+import type * as Api from "effect-http/Api";
 import type * as ClientError from "effect-http/ClientError";
 import type * as Route from "effect-http/Route";
-import * as ClientRequestEncoder from "effect-http/internal/clientRequestEncoder";
-import * as ClientResponseParser from "effect-http/internal/clientResponseParser";
-import * as Effect from "effect/Effect";
-import { identity, pipe } from "effect/Function";
-import * as Pipeable from "effect/Pipeable";
+import type * as Effect from "effect/Effect";
 import type * as Types from "effect/Types";
+
+import * as internal from "./internal/client";
 
 /**
  * @category models
@@ -24,46 +21,15 @@ import type * as Types from "effect/Types";
  */
 export type Client<Endpoints extends Api.Endpoint> = {
   [Id in Endpoints["id"]]: EndpointClient<Extract<Endpoints, { id: Id }>>;
-} & Pipeable.Pipeable;
-
-/** @internal */
-const httpClient = PlatformClient.fetch();
-
-/**
- * @category constructors
- * @since 1.0.0
- */
-export const endpointClient = <
-  Endpoints extends Api.Endpoint,
-  Id extends Endpoints["id"],
->(
-  id: Id,
-  api: Api.Api<Endpoints>,
-  options: Partial<Options>,
-): EndpointClient<Extract<Endpoints, { id: Id }>> => {
-  const endpoint = Api.getEndpoint(api, id);
-  const responseParser = ClientResponseParser.create(endpoint.schemas.response);
-  const requestEncoder = ClientRequestEncoder.create(endpoint);
-
-  let mapRequest = options.mapRequest ?? identity;
-
-  if (options.baseUrl) {
-    const url =
-      typeof options.baseUrl === "string"
-        ? options.baseUrl
-        : options.baseUrl.toString();
-    mapRequest = ClientRequest.prependUrl(url);
-  }
-
-  return (args: unknown) =>
-    pipe(
-      requestEncoder.encodeRequest(args),
-      Effect.map(mapRequest),
-      Effect.flatMap(httpClient),
-      Effect.flatMap(responseParser.parseResponse),
-      Effect.annotateLogs("clientOperationId", endpoint.id),
-    ) as any;
 };
+
+/** @ignore */
+export type EndpointClient<Endpoint extends Api.Endpoint> = ClientFunction<
+  Endpoint,
+  MakeHeadersOptionIfAllPartial<
+    Route.EndpointSchemasTo<Endpoint["schemas"]>["request"]
+  >
+>;
 
 /**
  * @category models
@@ -77,22 +43,28 @@ export interface Options {
 }
 
 /**
+ * @category constructors
+ * @since 1.0.0
+ */
+export const endpointClient: <
+  Endpoints extends Api.Endpoint,
+  Id extends Endpoints["id"],
+>(
+  id: Id,
+  api: Api.Api<Endpoints>,
+  options: Partial<Options>,
+) => EndpointClient<Extract<Endpoints, { id: Id }>> = internal.endpointClient;
+
+/**
  * Derive client implementation from the `Api`
  *
  * @category constructors
  * @since 1.0.0
  */
-export const make = <Endpoints extends Api.Endpoint>(
+export const make: <Endpoints extends Api.Endpoint>(
   api: Api.Api<Endpoints>,
   options?: Partial<Options>,
-): Client<Endpoints> =>
-  api.endpoints.reduce(
-    (client, endpoint) => ({
-      ...client,
-      [endpoint.id]: endpointClient(endpoint.id, api, options ?? {}),
-    }),
-    {} as Client<Endpoints>,
-  );
+) => Client<Endpoints> = internal.make;
 
 // Internal type helpers
 
@@ -138,11 +110,3 @@ type ClientFunction<Endpoint extends Api.Endpoint, I> = Record<
       ClientError.ClientError,
       ClientFunctionResponse<Endpoint["schemas"]["response"]>
     >;
-
-/** @ignore */
-type EndpointClient<Endpoint extends Api.Endpoint> = ClientFunction<
-  Endpoint,
-  MakeHeadersOptionIfAllPartial<
-    Route.EndpointSchemasTo<Endpoint["schemas"]>["request"]
-  >
->;
